Move website_analysis model into models
